+++ src/imio/dashboard/columns.py
@@ -1,4 +1,6 @@
 # encoding: utf-8
+
+from collective.contact.core.content.organization import IOrganization
 from collective.eeafaceted.z3ctable.columns import PrettyLinkColumn
 
 
@@ -15,7 +17,16 @@ class ContactPrettyLinkColumn(PrettyLinkColumn):
         """ """
         return getattr(item, self.attrName)()
 
-    def renderCell(self, item):
-        """ """
+    def getCSSClasses(self, item):
+        """Returns a CSS class specific to current content."""
+        cssClasses = super(ContactPrettyLinkColumn, self).getCSSClasses(item)
+
+        current_org_css_class = None
         obj = self._getObject(item)
-        return super(ContactPrettyLinkColumn, self).getPrettyLink(obj)
+        if IOrganization.providedBy(obj):
+            current_org_css_class = '___'.join([org.id for org in obj.get_organizations_chain()])
+
+        if current_org_css_class:
+            cssClasses['td'] = '{0} {1}'.format(cssClasses['td'], current_org_css_class)
+
+        return cssClasses